requirements.txt
# This file is used to install the requirements for the project. 
# Note that the versions are not given so that the latest version can be downloaded
# Run the below to install all the requirements
# pip install -r requirements.txt

moviepy
# If facing issues, run the following "sudo apt install ffmpeg"
ffmpeg
#below one takes a lot of time and space to install (1.5 Gb space and around 25 mins to install)
openai-whisper
datasets
transformers
scikit-learn #50MB
pydub
tensorboard
ffmpeg-python